% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_distributions.R
\name{capacity_cstar}
\alias{capacity_cstar}
\title{Critical capacity c* given current basin moisture S}
\usage{
capacity_cstar(S, dist, params)
}
\arguments{
\item{S}{Current soil moisture [mm].}

\item{dist}{One of \code{"pareto"}, \code{"rectangular"}, \code{"exponential"},
\code{"triangular"}, \code{"lognormal"}. These correspond directly to
Appendices A-E of Moore (2007).}

\item{params}{A \code{PdmParams} object or compatible named list.}
}
\value{
c* [mm].
}
\description{
Critical capacity c* given current basin moisture S
}
\references{
Moore (2007), Eq. 3 and Appendices A-F.
}
