% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_distributions.R
\name{capacity_smax}
\alias{capacity_smax}
\title{Maximum basin storage Smax for the chosen distribution}
\usage{
capacity_smax(dist, params)
}
\arguments{
\item{dist}{One of \code{"pareto"}, \code{"rectangular"}, \code{"exponential"},
\code{"triangular"}, \code{"lognormal"}. These correspond directly to
Appendices A-E of Moore (2007).}

\item{params}{A \code{PdmParams} object or compatible named list.}
}
\value{
Smax [mm].
}
\description{
Maximum basin storage Smax for the chosen distribution
}
\references{
Moore (2007), Appendices A-E.
}
