% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_io_compat.R
\name{.assert_hydrodata}
\alias{.assert_hydrodata}
\title{Assert that an object is a reach.io HydroData object}
\usage{
.assert_hydrodata(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{arg_name}{Name of the argument (used in error messages).}
}
\description{
Stops with an informative message if \code{x} is not a \code{HydroData} subclass,
or if reach.io is not installed.
}
\keyword{internal}
