% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_io_compat.R
\name{.assert_rainfall}
\alias{.assert_rainfall}
\title{Assert that a HydroData object is a Rainfall subclass}
\usage{
.assert_rainfall(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{HydroData object.}

\item{arg_name}{Argument name for error messages.}
}
\description{
Assert that a HydroData object is a Rainfall subclass
}
\keyword{internal}
