% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_stores.R
\name{.groundwater_step}
\alias{.groundwater_step}
\title{Update groundwater store one timestep}
\usage{
.groundwater_step(Sg, Qd, kb, m, Sg_max = 0)
}
\arguments{
\item{Sg}{Current groundwater store \verb{[mm]}.}

\item{Qd}{Recharge inflow this timestep \verb{[mm]}.}

\item{kb}{Baseflow time constant \verb{[hour * mm^\{1-m\}]} (Table 1: kb).}

\item{m}{Baseflow exponent (Table 1: m). m=1 gives linear reservoir.}

\item{Sg_max}{Max groundwater store \verb{[mm]}. 0 = unlimited.}
}
\value{
list(Sg = updated store \verb{[mm]}, Qb = baseflow \verb{[mm/timestep]})
}
\description{
Update groundwater store one timestep
}
\keyword{internal}
