% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_stores.R
\name{.surface_step}
\alias{.surface_step}
\title{Update surface routing cascade one timestep}
\usage{
.surface_step(Ss1, Ss2, Qs_in, k1, k2)
}
\arguments{
\item{Ss1}{State of first reservoir \verb{[mm]}.}

\item{Ss2}{State of second reservoir \verb{[mm]}.}

\item{Qs_in}{Surface runoff inflow this timestep \verb{[mm]}.}

\item{k1}{Time constant of first reservoir \verb{[hours]}.}

\item{k2}{Time constant of second reservoir \verb{[hours]}.}
}
\value{
list(Ss1, Ss2, Qf) where Qf is surface outflow \verb{[mm/timestep]}.
}
\description{
Implements the cascade of two linear reservoirs with time constants k1 and
k2 (Moore 2007, Table 1). Uses the exact recursive (Z-transform) solution.
}
\keyword{internal}
