% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh_fsr_uh.R
\name{feh_arf}
\alias{feh_arf}
\title{Areal reduction factor (FEH / FSR method)}
\usage{
feh_arf(area_km2, duration_hr, return_period = 10)
}
\arguments{
\item{area_km2}{Catchment area [km²].}

\item{duration_hr}{Storm duration [hours].}

\item{return_period}{Return period [years]. Affects ARF for T > 10 yr.}
}
\value{
ARF scalar in (0, 1].
}
\description{
Converts point rainfall to catchment-average rainfall for a given area,
duration, and return period using the FEH ARF regression equations
(Reed et al. 1999, eq 2.4).
}
