% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh_vol4_rainfall.R
\name{feh_ddf}
\alias{feh_ddf}
\title{Compute design rainfall depth using the FEH DDF model (FEH Vol 4)}
\usage{
feh_ddf(duration_hr, return_period, rmed_1h, rmed_1d, saar, region = "uk")
}
\arguments{
\item{duration_hr}{Storm duration [hours]. Scalar or vector.}

\item{return_period}{Return period [years]. Scalar or vector.
Must be compatible with \code{duration_hr} (recycled).}

\item{rmed_1h}{RMED for 1-hour duration [mm]. From FEH descriptors.}

\item{rmed_1d}{RMED for 1-day duration [mm]. From FEH descriptors.}

\item{saar}{Standard average annual rainfall [mm].}

\item{region}{FEH rainfall region. One of \code{"uk"} (default, uses
UK-wide parameters), or a named regional set.}
}
\value{
A \code{data.table} with columns \code{duration_hr}, \code{return_period},
\code{rainfall_mm}, \code{growth_factor}.
}
\description{
Estimates the T-year return period rainfall depth for a given duration
using the FEH depth-duration-frequency model. The site must be
characterised by its RMED (median annual maximum 1-hour rainfall) and
SAAR (standard average annual rainfall) catchment descriptors, available
from the FEH Web Service or NRFA catchment descriptors.
}
\examples{
# 100-year, 1-hour to 24-hour rainfall at a site with
# RMED(1h) = 12 mm, RMED(1d) = 38 mm, SAAR = 900 mm
feh_ddf(duration_hr = c(1, 2, 4, 6, 12, 24),
        return_period = 100,
        rmed_1h = 12, rmed_1d = 38, saar = 900)

}
\seealso{
\code{\link[=feh_design_storm]{feh_design_storm()}}, \code{\link[=refh2_run]{refh2_run()}}
}
