% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_frequency.R
\name{fit_gev}
\alias{fit_gev}
\title{Fit a GEV distribution to annual maxima (convenience wrapper)}
\usage{
fit_gev(annual_max, return_periods = c(2, 5, 10, 20, 50, 100, 200))
}
\arguments{
\item{annual_max}{Numeric vector of annual maximum flows.}

\item{return_periods}{Return periods \link{years}.}
}
\value{
A \code{FehStatFit} object. See \code{\link[=feh_single_site]{feh_single_site()}}.
}
\description{
Wrapper around \code{\link[=feh_single_site]{feh_single_site()}} with \code{dist = "gev"}. For the full
FEH-compliant analysis use \code{\link[=feh_single_site]{feh_single_site()}} directly.
}
\seealso{
\code{\link[=feh_single_site]{feh_single_site()}}, \code{\link[=fit_glo]{fit_glo()}}
}
