% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_frequency.R
\name{fit_glo}
\alias{fit_glo}
\title{Fit a GLO distribution to annual maxima (convenience wrapper)}
\usage{
fit_glo(annual_max, return_periods = c(2, 5, 10, 20, 50, 100, 200))
}
\arguments{
\item{annual_max}{Numeric vector of annual maximum flows.}

\item{return_periods}{Return periods \link{years}.}
}
\description{
Wrapper around \code{\link[=feh_single_site]{feh_single_site()}} with \code{dist = "glo"} (FEH-recommended
for GB). For the full FEH-compliant analysis use \code{\link[=feh_single_site]{feh_single_site()}}.
}
\seealso{
\code{\link[=feh_single_site]{feh_single_site()}}, \code{\link[=fit_gev]{fit_gev()}}
}
