% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_io_compat.R
\name{hydrodata_dates}
\alias{hydrodata_dates}
\title{Extract the date vector from a HydroData object}
\usage{
hydrodata_dates(x)
}
\arguments{
\item{x}{A reach.io HydroData object.}
}
\value{
Date vector.
}
\description{
Extract the date vector from a HydroData object
}
