% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_io_compat.R
\name{hydrodata_datetimes}
\alias{hydrodata_datetimes}
\title{Extract the dateTime vector from a HydroData object}
\usage{
hydrodata_datetimes(x)
}
\arguments{
\item{x}{A reach.io HydroData object.}
}
\value{
POSIXct vector.
}
\description{
Extract the dateTime vector from a HydroData object
}
