% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh_lmoments.R
\name{lmrd_theoretical}
\alias{lmrd_theoretical}
\title{L-moment ratio diagram data for common FEH distributions}
\usage{
lmrd_theoretical()
}
\value{
A \code{data.table} with columns \code{dist}, \code{t3}, \code{t4}.
}
\description{
L-moment ratio diagram data for common FEH distributions
}
