% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_metrics.R
\name{pbias}
\alias{pbias}
\title{Percent bias}
\usage{
pbias(obs, sim)
}
\arguments{
\item{obs}{Numeric vector of observed flow [any consistent unit].}

\item{sim}{Numeric vector of simulated flow, same length as \code{obs}.}
}
\value{
PBIAS [\%]. Positive = over-prediction; negative = under-prediction.
\code{NA} if sum of observed values is zero.
}
\description{
Percent bias
}
