% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh_lmoments.R
\name{qglo}
\alias{qglo}
\title{GLO quantile function}
\usage{
qglo(p, params)
}
\arguments{
\item{p}{Non-exceedance probability (or vector).}

\item{params}{A \code{GloParams} list (xi, alpha, k) or equivalent.}
}
\value{
Quantile(s) at probability p.
}
\description{
GLO quantile function
}
