% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feh_lmoments.R
\name{qgno}
\alias{qgno}
\title{GNO quantile function}
\usage{
qgno(p, params)
}
\arguments{
\item{p}{Non-exceedance probability (or vector).}

\item{params}{A \code{GloParams} list (xi, alpha, k) or equivalent.}
}
\description{
GNO quantile function
}
