% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdm_distributions.R
\name{capacity_smax}
\alias{capacity_smax}
\title{Maximum basin storage Smax for the chosen distribution}
\usage{
capacity_smax(dist, params)
}
\arguments{
\item{dist}{Distribution name. One of \code{"pareto"}, \code{"uniform"},
\code{"exponential"}, \code{"glogistic"}, \code{"normal"}, \code{"lognormal"}.}

\item{params}{A \code{PdmParams} object or compatible named list.}
}
\value{
Smax [mm].
}
\description{
Maximum basin storage Smax for the chosen distribution
}
\references{
Moore (2007), Appendices A-F.
}
